requirements.txt
# === Core ===
python-dotenv==1.0.1
tqdm==4.66.4
aiohttp==3.9.5
aiofiles==23.2.1
asyncio==3.4.3
requests==2.32.3
pandas==2.2.3
numpy==1.26.4

# === OpenAI / Embedding / LLM ===
openai>=1.30.0
tenacity==8.2.3

# === Text / NLP / Tokenization ===
nltk==3.9.1
spacy==3.7.5
transformers==4.43.3
sentencepiece==0.2.0
fugashi==1.3.2          # 日本語トークナイザ
unidic-lite==1.0.8      # 日本語辞書（軽量版）

# === Data Processing / File Handling ===
jsonlines==4.0.0
pyyaml==6.0.2
openpyxl==3.1.5
beautifulsoup4==4.12.3

# === Logging / Debugging ===
loguru==0.7.2
colorama==0.4.6
rich==13.7.1

# === Utilities ===
pathlib==1.0.1
regex==2024.4.28
typing-extensions==4.12.2

# === Optional (Visualization, only for inspection/debug) ===
matplotlib==3.9.2
seaborn==0.13.2
